+++ src/main/java/repository/datajpa/ProxyOrderRepository.java
@@ -12,15 +12,15 @@ import java.util.List;
  * Created by Александр on 14.11.2016.
  */
 @Transactional
-public interface ProxyOrderRepository extends JpaRepository<Order, Integer> {
+public interface ProxyOrderRepository extends JpaRepository<Order, Long> {
     @Override
-    Order getOne(Integer id);
+    Order getOne(Long id);
 
     @Override
     Order save(Order order);
 
     @Override
-    void delete(Integer id);
+    void delete(Long id);
 
     @Override
     List<Order> findAll();
