+++ src/main/resources/db/initDB.sql
@@ -18,7 +18,7 @@ CREATE SEQUENCE global_seq START 10;
 
 CREATE TABLE users
 (
-  id         INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id         int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   name       VARCHAR NOT NULL UNIQUE,
   email      VARCHAR NOT NULL UNIQUE,
   password   VARCHAR NOT NULL,
@@ -32,14 +32,14 @@ CREATE TABLE users
 
 CREATE TABLE user_roles
 (
-  user_id INTEGER NOT NULL,
+  user_id int8 NOT NULL,
   role    VARCHAR,
   CONSTRAINT user_roles_idx UNIQUE (user_id, role),
   FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
 );
 
 CREATE TABLE products (
-  id              INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id              int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   name            VARCHAR NOT NULL UNIQUE,
   description     TEXT,
   price           MONEY NOT NULL,
@@ -49,16 +49,16 @@ CREATE TABLE products (
 );
 
 CREATE TABLE orders (
-  id          INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id          int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   registered  TIMESTAMP DEFAULT now(),
-  user_id     integer REFERENCES users (id),
+  user_id     int8 REFERENCES users (id),
   paid        BOOL DEFAULT FALSE,
   statusOrder VARCHAR
 );
 
 CREATE TABLE order_products (
-  order_id    INTEGER NOT NULL,
-  product_id  INTEGER NOT NULL,
+  order_id    int8 NOT NULL,
+  product_id  int8 NOT NULL,
   number      INTEGER,
   UNIQUE (order_id, product_id),
   FOREIGN KEY (order_id) REFERENCES orders (id) ON DELETE CASCADE,
@@ -66,44 +66,44 @@ CREATE TABLE order_products (
 );
 
 CREATE TABLE devices (
-  id             INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
-  product_id     INTEGER REFERENCES products (id),
+  id             int8 PRIMARY KEY DEFAULT nextval('global_seq'),
+  product_id     int8 REFERENCES products (id),
   maxLed         INTEGER NOT NULL,
   enabled        BOOL DEFAULT FALSE,
   uuid           UUID,
-  user_id        INTEGER REFERENCES users (id)
+  user_id        int8 REFERENCES users (id)
 );
 
 CREATE TABLE audios(
-  id             INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id             int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   name           VARCHAR NOT NULL,
   path           VARCHAR NOT NULL
 );
 
 
 CREATE TABLE lightShows (
-  id             INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id             int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   name           VARCHAR NOT NULL,
-  device_id      INTEGER REFERENCES devices (id),
-  user_id        INTEGER REFERENCES users (id),
-  remix_user_id  INTEGER REFERENCES users (id),
+  device_id      int8 REFERENCES devices (id),
+  user_id        int8 REFERENCES users (id),
+  remix_user_id  int8 REFERENCES users (id),
   time           INTEGER NOT NULL,
-  audio_id       INTEGER REFERENCES audios (id)
+  audio_id       int8 REFERENCES audios (id)
 );
 
 CREATE TABLE effects(
-  id             INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id             int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   name           VARCHAR NOT NULL,
   commonTime     INTEGER NOT NULL,
   attenuation    BOOL DEFAULT FALSE,
   appearance     BOOL DEFAULT FALSE,
-  lightShow_id   INTEGER REFERENCES lightShows (id),
+  lightShow_id   int8 REFERENCES lightShows (id),
   FOREIGN KEY (lightShow_id) REFERENCES lightShows (id) ON DELETE CASCADE
 );
 
 CREATE TABLE effect_time_start (
-  lightShow_id   INTEGER REFERENCES lightShows (id),
-  effect_id      INTEGER REFERENCES effects (id),
+  lightShow_id   int8 REFERENCES lightShows (id),
+  effect_id      int8 REFERENCES effects (id),
   time           INTEGER NOT NULL,
   FOREIGN KEY (lightShow_id) REFERENCES lightShows (id) ON DELETE CASCADE,
   FOREIGN KEY (effect_id) REFERENCES effects (id) ON DELETE CASCADE
@@ -111,26 +111,26 @@ CREATE TABLE effect_time_start (
 
 
 CREATE TABLE leds(
-  id             INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
+  id             int8 PRIMARY KEY DEFAULT nextval('global_seq'),
   r              BYTEA,
   g              BYTEA,
   b              BYTEA,
   enabled        BOOL DEFAULT FALSE,
   number         INTEGER NOT NULL,
-  effect_id      INTEGER REFERENCES effects (id),
+  effect_id      int8 REFERENCES effects (id),
   FOREIGN KEY (effect_id) REFERENCES effects (id) ON DELETE CASCADE
 );
 
 
 CREATE TABLE effect_beginLedList(
-  effect_id      INTEGER REFERENCES effects (id),
-  led_id         INTEGER REFERENCES leds (id),
+  effect_id      int8 REFERENCES effects (id),
+  led_id         int8 REFERENCES leds (id),
   FOREIGN KEY (effect_id) REFERENCES effects (id) ON DELETE CASCADE
 );
 
 CREATE TABLE effect_endLedList(
-  effect_id      INTEGER REFERENCES effects (id),
-  led_id         INTEGER REFERENCES leds (id),
+  effect_id      int8 REFERENCES effects (id),
+  led_id         int8 REFERENCES leds (id),
   FOREIGN KEY (effect_id) REFERENCES effects (id) ON DELETE CASCADE
 );
